refactor: split stage into per-subcommand files and lift tarball helpers
The 631-line `stage.ts` carried six subcommands, tarball parsing,
auth/request plumbing, error class, OTP detection, and rendering helpers
in one file. Reorganized to match the existing `publish/` folder layout:

- `stage/{index,help,publish,list,view,approve,reject,download,context,
  request,parsing,rendering,errors,types}.ts` — one concept per file.
- `tarball/{publishSummary,summarizeTarball}.ts` — shared between
  `publish` and `stage` instead of duplicated. `PublishSummary` and
  `extractBundledDependencies` now live with the tarball helper rather
  than inside the publish subfolder, so other commands can reuse them
  without reaching into `publish/`.

Behavior unchanged. Also dropped `StageRegistryError`'s redundant
`statusCode` field (was identical to `status`) to bring it in line with
`FailedToPublishError`.
