+++ releasing/commands/src/index.ts
@@ -1,5 +1,5 @@
 export { deploy } from './deploy/index.js'
 export { packApp } from './pack-app/index.js'
 export { pack, publish } from './publish/index.js'
-export * as stage from './stage.js'
+export * as stage from './stage/index.js'
 export { version } from './version/index.js'
