+++ releasing/commands/src/tarball/index.ts
@@ -1,0 +1,2 @@
+export { extractBundledDependencies, type PublishSummary } from './publishSummary.js'
+export { summarizeTarball } from './summarizeTarball.js'
